chore: Refactor hashData and useMulticall functions
