+++ src/lib/multicall/multicall.worker.ts
@@ -10,8 +10,6 @@ async function executeMulticall(chainId: number, request: MulticallRequestConfig
 self.addEventListener("message", run);
 
 async function run(event) {
-  performance.mark("multicall-worker-start");
-  const start = Date.now();
   const { PRODUCTION_PREVIEW_KEY, chainId, request, id } = event.data;
   // @ts-ignore
   self.PRODUCTION_PREVIEW_KEY = PRODUCTION_PREVIEW_KEY;
@@ -26,13 +24,9 @@ async function run(event) {
   } catch (error) {
     postMessage({ id, error: error });
   }
-
-  const end = Date.now();
-  console.log("multicall-worker", end - start, "ms");
-  performance.mark("multicall-worker-end");
-  performance.measure("multicall-worker", "multicall-worker-start", "multicall-worker-end");
 }
 
+// Typescript hack to make it seem this file exports a class
 declare class MulticallWorker extends Worker {
   constructor();
 }
